+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.36.3
+
+## Component Enhancements
+
+* `DialogFullScreen` now accepts a String for title or any other component.
+
 # 0.36.2
 
 * Hide SplitButton additional buttons instead of removing them.
@@ -35,7 +41,7 @@ with a label.
 </InlineInputs>
 ```
 
-## Component Enchancements
+## Component Enhancements
 
 * `Date` now shows error validation when an invalid date is entered.
 * `Flash`: Change error icon to match other notifications (now shows error icon when `as` prop is `error`)